[#71797] Fix `add-missing-brda-entries`
Previously line numbers were stored as strings so passing string was
correct even though according to the typing it should be int.

In d74328af the way we store line information was modified and we've
been storing ints since then instead. It broke the
`add-missing-brda-entries` option because the result of
`has_entry_for_line` was always False.